update: change code for updating to app router
