+++ src/createEmotionCache.js
@@ -1,5 +1,5 @@
-import createCache from "@emotion/cache";
+import createCache from '@emotion/cache';
 
 export default function createEmotionCache() {
-  return createCache({ key: "css", prepend: true });
+  return createCache({ key: 'css', prepend: true });
 }
